requirements.txt
flask==3.0.0
anthropic==0.40.0
openai-whisper
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
click==8.1.7
yt-dlp